requirements.txt
django==1.9.2
django-gcm==1.2.0
django-rest-framework=3.6.2
psycopg2==2.6.2
gunicorn==19.6.0
openpyxl==2.3.5
mt-940=4.5
django-iban-field==0.8
mt-940==4.5